requirements.txt
# RoluATM Python Backend Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Serial Communication (for T-Flex dispenser)
pyserial==3.5

# Environment Configuration
python-dotenv==1.0.0

# Crypto Libraries (if needed)
cryptography==41.0.8

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Logging and Monitoring
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production WSGI Server
gunicorn==21.2.0

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3

# Optional: For production monitoring
# prometheus-client==0.19.0
# sentry-sdk==1.39.2

# Hardware Testing Tools
# RPi.GPIO==0.7.1  # Uncomment for actual Raspberry Pi GPIO control 